move Sandbox TreeItem types to a dedicated file
